pyproject.toml
[build-system]
requires = ["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-cli"
version = "0.1.0"
description = "A CLI to manage AI tools and resources"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "rich>=12.0.0",
    "GitPython>=3.1.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "pytest-mock>=3.10.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
ai-cli = "ai_cli.main:main"

[tool.setuptools]
# Explicitly specify the packages to include
packages = ["ai_cli", "prompts"]

# Explicitly list all Python modules (files) to include
py-modules = []

# Include all data files in the package directories
include-package-data = true

[tool.setuptools.package-data]
ai_cli = ["*.txt", "*.json", "*.yaml", "*.yml"]
prompts = ["*.txt", "*.json", "*.yaml", "*.yml"]

# Explicitly exclude any unwanted files from all packages
[tool.setuptools.exclude-package-data]
"*" = ["*.pyc", "*.pyo", "*~", "*.bak", "*.swp"]
